+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 * Sort comments by descending net score (breaks BC).
 
+## v0.1.2 (201x-xx-xx)
+
+* Make use of Doctrine migrations.
+* Add missing 'create forum' link in the menu on the front page.
+* Add a form theme and CSS so all forms look OK.
+
 ## v0.1.1 (2016-12-29)
 
 * Added the ability for the user to choose how to sort submission listings.